refector: 로직 변경, 로그 일부 제거, 관련 일정/할일/금액 id 추가
